Implemented and greatly optimized friend request functionality

+++ src/Controller/GameController.php
@@ -3,12 +3,16 @@
 namespace App\Controller;
 
 use App\Entity\Game;
+use App\Entity\User;
 use App\Form\GameType;
+use DateTimeImmutable;
 use App\Repository\GameRepository;
 use App\Repository\GameResultsRepository;
 use CMEN\GoogleChartsBundle\GoogleCharts\Charts\LineChart;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -30,15 +34,37 @@ class GameController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+			/** @var UploadedFile $imageFile */
+			$imageFile = $form->get('image')->getData();
+			if ($imageFile) {
+				$filename = uniqid().'.'.$imageFile->guessExtension();
+				try {
+					$imageFile->move(
+						'assets/images/games/',
+						$filename
+					);
+                } catch (FileException $e) {
+					// ... handle exception if something happens during file upload
+				}
+				$game->setImage('assets/images/games/' . $filename);
+			}
+			// set developer - important!!
+			$game->setDeveloper($this->getUser());
+			// generate API key
+			$game->setApiKey(uniqid("", true));
+			// also store creation time
+			$game->setCreatedOn(new DateTimeImmutable('now'));
+			$game->setUpdatedOn(new DateTimeImmutable('now'));
+			// persist & flush data
             $entityManager->persist($game);
             $entityManager->flush();
-
-            return $this->redirectToRoute('app_game_index', [], Response::HTTP_SEE_OTHER);
+			// redirect to game page
+			return $this->redirectToRoute('app_game_show', ['id' => $game->getId()], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('game/new.html.twig', [
             'game' => $game,
-            'form' => $form,
+            'gameForm' => $form,
         ]);
     }
 
@@ -46,6 +72,14 @@ class GameController extends AbstractController
 	#[Route('/{id}/leaderboard', name: 'app_game_show_results', methods: ['GET'])]
     public function show(Game $game) : Response
     {
+		/** @var $appUser User */
+		$appUser = $this->getUser();
+		if (!$game->isPublic() &&
+				!$appUser->isAdmin() &&
+				$appUser != $game->getDeveloper() &&
+				!in_array($appUser, $game->getDeveloper()->getFriends()->toArray())) {
+			throw $this->createAccessDeniedException('You are not authorized to access this game!');
+		}
 		$sort = 'score';
 		if (isset($_GET['sort'])) {
 			$sort = htmlspecialchars($_GET['sort']);
@@ -61,6 +95,14 @@ class GameController extends AbstractController
 	#[Route('/{id}/stats', name: 'app_game_show_stats', methods: ['GET'])]
     public function show_stats(Game $game): Response
     {
+		/** @var $appUser User */
+		$appUser = $this->getUser();
+		if (!$game->isPublic() &&
+				!$appUser->isAdmin() &&
+				$appUser != $game->getDeveloper() &&
+				!in_array($appUser, $game->getDeveloper()->getFriends()->toArray())) {
+			throw $this->createAccessDeniedException('You are not authorized to access this game!');
+		}
 		$scoreChartData = [
 			['Time', 'Score', ['role' => 'annotation']],
 			[$game->getCreatedOn(), 0, null],
@@ -84,25 +126,56 @@ class GameController extends AbstractController
     #[Route('/{id}/edit', name: 'app_game_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Game $game, EntityManagerInterface $entityManager): Response
     {
+		/** @var $appUser User */
+		$appUser = $this->getUser();
+		if ($appUser != $game->getDeveloper() && !$appUser->isAdmin() ) {
+			throw $this->createAccessDeniedException('You are not authorized to edit this game!');
+		}
+
         $form = $this->createForm(GameType::class, $game);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+			/** @var UploadedFile $imageFile */
+			$imageFile = $form->get('image')->getData();
+			if ($imageFile) {
+				unlink($game->getImage());
+				$filename = uniqid().'.'.$imageFile->guessExtension();
+				try {
+					$imageFile->move(
+						'assets/images/games/',
+						$filename
+					);
+                } catch (FileException $e) {
+					// ... handle exception if something happens during file upload
+				}
+				$game->setImage('assets/images/games/' . $filename);
+			}
+			// also store time of update
+			$game->setUpdatedOn(new DateTimeImmutable('now'));
+			// flush data
             $entityManager->flush();
-
-            return $this->redirectToRoute('app_game_index', [], Response::HTTP_SEE_OTHER);
+			// redirect to game page
+			return $this->redirectToRoute('app_game_show', ['id' => $game->getId()], Response::HTTP_SEE_OTHER);
         }
 
         return $this->render('game/edit.html.twig', [
             'game' => $game,
-            'form' => $form,
+            'gameForm' => $form->createView(),
         ]);
     }
 
     #[Route('/{id}', name: 'app_game_delete', methods: ['POST'])]
     public function delete(Request $request, Game $game, EntityManagerInterface $entityManager): Response
     {
+		/** @var $appUser User */
+		$appUser = $this->getUser();
+		if ($appUser != $game->getDeveloper() && !$appUser->isAdmin() ) {
+			throw $this->createAccessDeniedException('You are not authorized to delete this game!');
+		}
+
         if ($this->isCsrfTokenValid('delete'.$game->getId(), $request->request->get('_token'))) {
+			unlink($game->getImage());
             $entityManager->remove($game);
             $entityManager->flush();
         }
